refactor: preprocessing
